ai/requirements.txt
google-generativeai
pypdf2
requests
beautifulsoup4
python-dotenv